+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/internal/JavaLanguageProcessor.java
@@ -36,6 +36,7 @@ import net.sourceforge.pmd.reporting.ViolationSuppressor;
 import net.sourceforge.pmd.util.AuxClasspathLoader;
 import net.sourceforge.pmd.util.designerbindings.DesignerBindings;
 import net.sourceforge.pmd.util.internal.AuxClasspathUtil;
+import net.sourceforge.pmd.util.log.internal.LogUtil;
 
 /**
  * @author Clément Fournier
@@ -72,15 +73,21 @@ public class JavaLanguageProcessor extends BatchLanguageProcessor<JavaLanguagePr
             String rawAuxClasspath = properties.getProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH);
             List<Path> auxClasspath = new ArrayList<>(AuxClasspathUtil.expandClasspath(rawAuxClasspath));
 
+            Boolean disableWarnings = properties.getProperty(JavaLanguageProperties.DISABLE_AUX_CLASSPATH_WARNINGS);
+            LogUtil.WarnOrDebugLogger warnOrDebugLogger = LogUtil.createWarnOrDebugLogger(!disableWarnings,
+                    "Set env var PMD_JAVA_DISABLE_AUX_CLASSPATH_WARNINGS=true to disable this warning.");
             if (!AuxClasspathUtil.containsPlatformClasspath(auxClasspath)) {
                 Path platformClasspath = AuxClasspathUtil.getPlatformClasspath();
-                LOG.warn("Adding current platform {} to auxClasspath, which could be the wrong java version. "
-                        + "Please add the correct jrt-fs.jar explicitly to the auxClasspath.", platformClasspath);
+
+                warnOrDebugLogger.log(LOG, "Adding current platform {} to auxClasspath, which could be the wrong java version. "
+                                + "Please add the correct jrt-fs.jar explicitly to the auxClasspath. "
+                                + "See https://docs.pmd-code.org/latest/pmd_languages_java.html#providing-the-auxiliary-classpath",
+                        platformClasspath);
                 auxClasspath.add(platformClasspath);
             }
             String expandedAuxClasspath = AuxClasspathUtil.toRawClasspath(auxClasspath);
             LOG.debug("Using auxClasspath as analysis classloader: {}", expandedAuxClasspath);
-            this.auxClasspathLoader = AuxClasspathLoader.create(expandedAuxClasspath);
+            this.auxClasspathLoader = AuxClasspathLoader.create(expandedAuxClasspath, warnOrDebugLogger);
             this.typeSystem = TypeSystem.usingClasspath(name -> auxClasspathLoader.findResource(name));
         }
     }