+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/internal/JavaLanguageProcessorTest.java
@@ -6,6 +6,8 @@ package net.sourceforge.pmd.lang.java.internal;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.emptyString;
+import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -13,6 +15,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -20,19 +23,61 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 
 import net.sourceforge.pmd.lang.JvmLanguagePropertyBundle;
+import net.sourceforge.pmd.lang.LanguageProcessorRegistry;
 import net.sourceforge.pmd.lang.LanguagePropertyBundle;
+import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.lang.java.JavaLanguageModule;
 import net.sourceforge.pmd.lang.java.symbols.JClassSymbol;
+import net.sourceforge.pmd.util.CollectionUtil;
 import net.sourceforge.pmd.util.internal.AuxClasspathUtil;
+import net.sourceforge.pmd.util.log.PmdReporter;
 
 import com.github.stefanbirkner.systemlambda.SystemLambda;
+import uk.org.webcompere.systemstubs.SystemStubs;
+import uk.org.webcompere.systemstubs.environment.EnvironmentVariables;
+import uk.org.webcompere.systemstubs.jupiter.SystemStubsExtension;
 
+@ExtendWith(SystemStubsExtension.class)
 class JavaLanguageProcessorTest {
+
     @TempDir
     private Path tempDir;
+
+    @Test
+    void expectAuxClasspathWarning() throws Exception {
+        String classpath = AuxClasspathUtil.toRawClasspath(AuxClasspathUtil.getRuntimeClasspath());
+        JavaLanguageProperties properties =
+                (JavaLanguageProperties) JavaLanguageModule.getInstance().newPropertyBundle();
+        properties.setProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH, classpath);
+        String log = SystemStubs.tapSystemErr(() -> {
+            try (JavaLanguageProcessor processor = new JavaLanguageProcessor(properties)) {
+                assertNotNull(processor.getTypeSystem());
+            }
+        });
+        assertThat(log, containsString("Adding current platform"));
+    }
+
+    @Test
+    void expectWarningForCorruptJar() throws Exception {
+        Path corruptJar = tempDir.resolve("corrupt.jar");
+        Files.write(corruptJar, "PK\003\004 Corrupt Jar".getBytes(StandardCharsets.US_ASCII));
+        String classpath = AuxClasspathUtil.toRawClasspath(CollectionUtil.listOf(corruptJar), AuxClasspathUtil.getPlatformClasspath());
+        JavaLanguageProperties properties =
+                (JavaLanguageProperties) JavaLanguageModule.getInstance().newPropertyBundle();
+        properties.setProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH, classpath);
+        String log = SystemStubs.tapSystemErr(() -> {
+            try (JavaLanguageProcessor processor = new JavaLanguageProcessor(properties)) {
+                assertNotNull(processor.getTypeSystem());
+            }
+        });
+        assertThat(log, containsString("Ignoring corrupt archive on auxClasspath"));
+        assertThat(log, containsString(corruptJar.toString()));
+        assertThat(log, containsString("PMD_JAVA_DISABLE_AUX_CLASSPATH_WARNINGS"));
+    }
 
     @Test
     void classpathListWithJrtFs() throws Exception {
@@ -46,6 +91,44 @@ class JavaLanguageProcessorTest {
         String auxClasspath = writeClasspathFile("classpath-without-jrtfs.txt", false);
         String log = SystemLambda.tapSystemErr(() -> assertPlatformClassesFound(auxClasspath));
         assertThat(log, containsString("Adding current platform"));
+    }
+
+    @Test
+    void expectNoAuxClasspathWarningViaLanguageProperty() throws Exception {
+        String classpath = AuxClasspathUtil.toRawClasspath(AuxClasspathUtil.getRuntimeClasspath());
+        JavaLanguageProperties properties =
+                (JavaLanguageProperties) JavaLanguageModule.getInstance().newPropertyBundle();
+        properties.setProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH, classpath);
+        properties.setProperty(JavaLanguageProperties.DISABLE_AUX_CLASSPATH_WARNINGS, true);
+        String log = SystemStubs.tapSystemErr(() -> {
+            try (JavaLanguageProcessor processor = new JavaLanguageProcessor(properties)) {
+                assertNotNull(processor.getTypeSystem());
+            }
+        });
+        assertThat(log, is(emptyString()));
+    }
+
+    @Test
+    void expectNoAuxClasspathWarningViaEnvironmentVariable(EnvironmentVariables environment) throws Exception {
+        environment.set("PMD_JAVA_DISABLE_AUX_CLASSPATH_WARNINGS", "true");
+
+        String classpath = AuxClasspathUtil.toRawClasspath(AuxClasspathUtil.getRuntimeClasspath());
+        JavaLanguageModule javaLanguageModule = JavaLanguageModule.getInstance();
+        JavaLanguageProperties properties = (JavaLanguageProperties) javaLanguageModule.newPropertyBundle();
+        properties.setProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH, classpath);
+
+        try (LanguageProcessorRegistry registry =
+                     LanguageProcessorRegistry.create(LanguageRegistry.singleton(javaLanguageModule),
+                CollectionUtil.mapOf(javaLanguageModule, properties),
+                PmdReporter.quiet())) {
+            String log = SystemStubs.tapSystemErr(() -> {
+                try (JavaLanguageProcessor processor =
+                             (JavaLanguageProcessor) registry.getProcessor(javaLanguageModule)) {
+                    assertNotNull(processor.getTypeSystem());
+                }
+            });
+            assertThat(log, is(emptyString()));
+        }
     }
 
     @Test